refactor(content): split body.go into per-format files
body.go had accumulated 980 LOC across four release cycles —
PDF, four office formats, EPUB, and email (eml + mbox) — plus
shared XML/HTML text-strip helpers. It was the densest production
file in the codebase and the lone multi-format family without a
per-file split.

Mirrors the existing convention (disk_*.go, install_*.go,
bytecode_*.go, audio_*.go, video_*.go, binary_*.go, archive_*.go,
email_*.go) by moving each format's extractor to its own file:

  body.go              51 LOC  ExtractBody dispatcher only
  body_shared.go      206 LOC  extractXMLText, extractHTMLText,
                               htmlAutoClose / htmlEntities /
                               htmlBlockElems, endsWithNewline
  body_pdf.go         141 LOC  pdfBody + pdfCIDPattern
  body_office.go      245 LOC  ooxmlBody, xlsxBody, pptxBody,
                               odtBody
  body_epub.go        123 LOC  epubBody, readEPUBSpineHrefs
  body_email.go       254 LOC  emlBody, emlBodyFromReader,
                               mboxBody, walkEmailBody,
                               walkMultipartBody, decodeTextPart

Pure refactor — no behaviour change, no API change. ExtractBody's
signature is identical, the sole external caller
(internal/celexpr/body.go) is untouched, and TestExtractBody_Fixtures
exercises all eight format paths against committed fixtures.

Verification: go build, go vet, go fix -diff (empty),
go test -race ./..., golangci-lint run — all clean.
